requirements.txt
flask==3.0.2
python-dotenv==1.0.1
pymongo==4.6.1
python-docx==1.1.0
PyPDF2==3.0.1
flask-cors==4.0.0
flask-restx==1.3.0
python-magic==0.4.27 